refactor: extract nibble exception into a separate package
