+++ src/main/java/com/limechain/trie/structure/nibble/exceptions/NibbleException.java
@@ -1,0 +1,15 @@
+package com.limechain.trie.structure.nibble.exceptions;
+
+public class NibbleException extends RuntimeException {
+    public NibbleException(Throwable cause) {
+        super(cause);
+    }
+
+    public NibbleException(String message) {
+        super(message);
+    }
+
+    public NibbleException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
